Fixed tests without changing old data.

+++ part4/bloglist/controllers/blogs.js
@@ -44,10 +44,12 @@ blogsRouter.delete('/:id', async (request, response) => {
     }
 
     const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(400).json({ error: 'invalid blog id' })
+    }
     if (!(blog.user.toString() === decodedToken.id.toString())) {
         return response.status(401).json({ error: 'access denied: invalid user' })
     }
-
     await Blog.deleteOne(blog)
     response.status(204).end()
 })
